Write custom printf implementation
I'm trying to work on #2, this implementation will allow us to move to a
"write" syscall, instead of this putchar stuff we currently have.

(processes are really more like tasks at the moment and share all global
variables, the old printf implementation would make it really difficult
to buffer nicely).